+++ processor/src/test/resources/bad_input/com/example/computed/BadStreamTypeModel.java
@@ -1,0 +1,15 @@
+package com.example.computed;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Computed;
+import java.util.stream.Stream;
+
+@ArezComponent
+public abstract class BadStreamTypeModel
+{
+  @Computed
+  public Stream<Integer> setField()
+  {
+    return null;
+  }
+}