+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   so that the values are wrapped as unmodifiable variants if `Arez.areCollectionsPropertiesUnmodifiable()`
   returns true. The goal is to disallow the scenario where an observer accesses and accidentally or
   deliberately modifies the collection as changes will not be propagated to other observers of the property.
+* **\[processor\]** Generate an error if a `@Computed` method has a return type of `java.util.stream.Stream`
+  as instances of the type are single use and not candidates for caching.
 
 ### [v0.85](https://github.com/arez/arez/tree/v0.85) (2018-05-23)
 [Full Changelog](https://github.com/arez/arez/compare/v0.84...v0.85)